Implement document editing as user control  (WIP, tabs detection still not working)
