+++ R7.University.Launchpad/EditEduProgram.ascx.cs
@@ -11,28 +11,11 @@ using R7.University.ControlExtensions;
 
 namespace R7.University.Launchpad
 {
+    // available tabs
+    public enum EditEduProgramTab { Common, Documents };
+
 	public partial class EditEduProgram : LaunchpadPortalModuleBase
 	{
-        #region Types
-
-        public enum EditEduProgramTab { Common, Documents };
-
-        #endregion
-
-        #region Bindable icons
-
-        protected string EditIconUrl
-        {
-            get { return IconController.IconURL ("Edit"); }
-        }
-
-        protected string DeleteIconUrl
-        {
-            get { return IconController.IconURL ("Delete"); }
-        }
-
-        #endregion
-
         protected EditEduProgramTab SelectedTab
         {
             get 
@@ -41,7 +24,7 @@ namespace R7.University.Launchpad
                 var eventTarget = Request.Form ["__EVENTTARGET"];
 
                 // document URL control is on Documents tab
-                if (!string.IsNullOrEmpty (eventTarget) && eventTarget.Contains ("$" + urlDocumentUrl.ID +"$"))
+                if (!string.IsNullOrEmpty (eventTarget) && eventTarget.Contains ("$urlDocumentUrl$"))
                 {
                     ViewState ["SelectedTab"] = EditEduProgramTab.Documents;
                     return EditEduProgramTab.Documents;
@@ -63,43 +46,7 @@ namespace R7.University.Launchpad
 		// ALT: private int itemId = Null.NullInteger;
 		private int? itemId = null;
 
-        #region Documents & document types
-
-        protected void RememberDocumentTypes (IEnumerable<DocumentTypeInfo> documentTypes)
-        {
-            ViewState ["documentTypes"] = documentTypes.ToList ();
-        }
-
-        protected string GetDocumentType (int? documentTypeId)
-        {
-            if (documentTypeId != null)
-            {
-                var documentTypes = (List<DocumentTypeInfo>) ViewState ["documentTypes"];
-                return documentTypes.Single (dt => dt.DocumentTypeID == documentTypeId.Value).Type;
-            }
-
-            return string.Empty;
-        }
-
-        private List<DocumentInfo> GetDocuments ()
-        {
-            if (ViewStateDocuments != null)
-            {
-                return ViewStateDocuments.Select (dvm => dvm.NewDocumentInfo ()).ToList ();
-            }
-
-            return new List<DocumentInfo> ();
-        }
-
-        protected List<DocumentViewModel> ViewStateDocuments
-        {
-            get { return XmlSerializationHelper.Deserialize<List<DocumentViewModel>> (ViewState ["documents"]); }
-            set { ViewState ["documents"] = XmlSerializationHelper.Serialize<List<DocumentViewModel>> (value); }
-        }
-
-        #endregion
-
-		#region Handlers
+		#region Overrides
 
 		/// <summary>
 		/// Handles Init event for a control.
@@ -113,20 +60,15 @@ namespace R7.University.Launchpad
 			linkCancel.NavigateUrl = Globals.NavigateURL ();
 
 			// add confirmation dialog to delete button
-			buttonDelete.Attributes.Add ("onClick", "javascript:return confirm('" + Localization.GetString ("DeleteItem") + "');");
+			buttonDelete.Attributes.Add ("onClick", "javascript:return confirm('" 
+                + Localization.GetString ("DeleteItem") + "');");
 
 			// bind education levels
             comboEduLevel.DataSource = LaunchpadController.GetObjects<EduLevelInfo> ();
             comboEduLevel.DataBind ();
 
-            // bind document types
             var documentTypes = LaunchpadController.GetObjects<DocumentTypeInfo> ();
-            RememberDocumentTypes (documentTypes);
-
-            comboDocumentType.DataSource = DocumentTypeViewModel.GetBindableList (documentTypes, LocalResourceFile, true);
-            comboDocumentType.DataBind ();
-           
-            gridDocuments.LocalizeColumns (LocalResourceFile);
+            formEditDocuments.OnInit (this, documentTypes);
 		}
 
 		/// <summary>
@@ -169,11 +111,8 @@ namespace R7.University.Launchpad
                             auditControl.Bind (item);
 
                             var documents = LaunchpadController.GetObjects<DocumentInfoEx> (string.Format (
-                                "WHERE ItemID = N'EduProgramID={0}'", item.EduProgramID)).Select (d => new DocumentViewModel (d, ViewModelContext)).ToList ();
-
-                            ViewStateDocuments = documents;
-                            gridDocuments.DataSource = DataTableConstructor.FromIEnumerable (documents);
-                            gridDocuments.DataBind ();
+                                "WHERE ItemID = N'EduProgramID={0}'", item.EduProgramID)).ToList ();
+                            formEditDocuments.SetDocuments (item.EduProgramID, documents);
 						}
 						else
 							Response.Redirect (Globals.NavigateURL (), true);
@@ -190,6 +129,20 @@ namespace R7.University.Launchpad
 				Exceptions.ProcessModuleLoadException (this, ex);
 			}
 		}
+
+        protected override void OnUnload (EventArgs e)
+        {
+            base.OnUnload (e);
+
+            if (formEditDocuments.IsActive)
+            {
+                SelectedTab = EditEduProgramTab.Documents;
+            }
+        }
+
+        #endregion
+
+        #region Handlers
 
 		/// <summary>
 		/// Handles Click event for Update button
@@ -234,7 +187,7 @@ namespace R7.University.Launchpad
                     item.LastModifiedOnDate = item.CreatedOnDate;
                     item.CreatedByUserID = UserInfo.UserID;
                     item.LastModifiedByUserID = item.CreatedByUserID;
-                    LaunchpadController.AddEduProgram (item, GetDocuments ());
+                    LaunchpadController.AddEduProgram (item, formEditDocuments.GetDocuments ());
                 }
 				else
                 {
@@ -248,7 +201,7 @@ namespace R7.University.Launchpad
                         item.CreatedByUserID = item.LastModifiedByUserID;
                     }
 
-                    LaunchpadController.UpdateEduProgram (item, GetDocuments ());
+                    LaunchpadController.UpdateEduProgram (item, formEditDocuments.GetDocuments ());
                 }
 
 				Utils.SynchronizeModule (this);
@@ -287,194 +240,6 @@ namespace R7.University.Launchpad
 		}
 
 		#endregion
-       
-        #region Documents
-
-        protected void buttonAddDocument_Command (object sender, CommandEventArgs e)
-        {
-            try
-            {
-                SelectedTab = EditEduProgramTab.Documents;
-
-                DocumentViewModel document;
-
-                // get documents list from viewstate
-                var documents = ViewStateDocuments;
-
-                // creating new list, if none
-                if (documents == null)
-                    documents = new List<DocumentViewModel> ();
-
-                var command = e.CommandArgument.ToString ();
-                if (command == "Add")
-                {
-                    document = new DocumentViewModel ();
-                }
-                else
-                {
-                    // restore ItemID from hidden field
-                    var hiddenItemID = int.Parse (hiddenDocumentItemID.Value);
-                    document = documents.Find (d => d.ViewItemID == hiddenItemID);
-                }
-
-                document.Title = textDocumentTitle.Text.Trim ();
-                document.DocumentTypeID = Utils.ParseToNullableInt (comboDocumentType.SelectedValue);
-                document.Type = GetDocumentType (document.DocumentTypeID);
-                document.SortIndex = int.Parse (textDocumentSortIndex.Text);
-                document.StartDate = datetimeDocumentStartDate.SelectedDate;
-                document.EndDate = datetimeDocumentEndDate.SelectedDate;
-                document.Url = urlDocumentUrl.Url;
-
-                if (command == "Add")
-                {
-                    document.ItemID = "EduProgramID=" + itemId.Value;
-                    documents.Add (document);
-                }
-
-                ResetEditDocumentForm ();
-
-                // refresh viewstate
-                ViewStateDocuments = documents;
-
-                // bind items to the gridview
-                DocumentViewModel.BindToView (documents, ViewModelContext);
-                gridDocuments.DataSource = DataTableConstructor.FromIEnumerable (documents);
-                gridDocuments.DataBind ();
-                
-            }
-            catch (Exception ex)
-            {
-                Exceptions.ProcessModuleLoadException (this, ex);
-            }
-        }
-
-        protected void linkEditDocument_Command (object sender, CommandEventArgs e)
-        {
-            try
-            {
-                SelectedTab = EditEduProgramTab.Documents;
-
-                var documents = ViewStateDocuments;
-                if (documents != null)
-                {
-                    var itemID = e.CommandArgument.ToString ();
-
-                    // find document in a list
-                    var document = documents.Find (d => d.ViewItemID.ToString () == itemID);
-
-                    if (document != null)
-                    {
-                        // fill form
-                        Utils.SelectByValue (comboDocumentType, document.DocumentTypeID);
-                        textDocumentTitle.Text = document.Title;
-                        textDocumentSortIndex.Text = document.SortIndex.ToString ();
-                        datetimeDocumentStartDate.SelectedDate = document.StartDate;
-                        datetimeDocumentEndDate.SelectedDate = document.EndDate;
-                        urlDocumentUrl.Url = document.Url;
-
-                        // store ItemID in the hidden field
-                        hiddenDocumentItemID.Value = document.ViewItemID.ToString ();
-
-                        // show / hide buttons
-                        buttonAddDocument.Visible = false;
-                        buttonUpdateDocument.Visible = true;
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                Exceptions.ProcessModuleLoadException (this, ex);
-            }
-        }
-
-        protected void linkDeleteDocument_Command (object sender, CommandEventArgs e)
-        {
-            try
-            {
-                SelectedTab = EditEduProgramTab.Documents;
-
-                var documents = ViewStateDocuments;
-                if (documents != null)
-                {
-                    var itemID = e.CommandArgument.ToString ();
-
-                    // find position in a list
-                    var documentIndex = documents.FindIndex (d => d.ViewItemID.ToString () == itemID);
-
-                    if (documentIndex >= 0)
-                    {
-                        // remove item
-                        documents.RemoveAt (documentIndex);
-
-                        // refresh viewstate
-                        ViewStateDocuments = documents;
-
-                        // bind to the gridview
-                        DocumentViewModel.BindToView (documents, ViewModelContext);
-                        gridDocuments.DataSource = DataTableConstructor.FromIEnumerable (documents);
-                        gridDocuments.DataBind ();
-
-                        // reset form if we deleting currently edited item
-                        if (buttonUpdateDocument.Visible && hiddenDocumentItemID.Value == itemID)
-                            ResetEditDocumentForm ();
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                Exceptions.ProcessModuleLoadException (this, ex);
-            }
-        }
-
-        protected void buttonCancelEditDocument_Click (object sender, EventArgs e)
-        {
-            try
-            {
-                SelectedTab = EditEduProgramTab.Documents;
-
-                ResetEditDocumentForm ();
-            }
-            catch (Exception ex)
-            {
-                Exceptions.ProcessModuleLoadException (this, ex);
-            }
-        }
-
-        void ResetEditDocumentForm ()
-        {
-            // restore default buttons visibility
-            buttonAddDocument.Visible = true;
-            buttonUpdateDocument.Visible = false;
-
-            comboDocumentType.SelectedIndex = 0;
-            textDocumentTitle.Text = string.Empty;
-            textDocumentSortIndex.Text = "0";
-            datetimeDocumentStartDate.SelectedDate = null;
-            datetimeDocumentEndDate.SelectedDate = null;
-            urlDocumentUrl.UrlType = "N";
-        }
-
-        protected void gridDocuments_RowDataBound (object sender, GridViewRowEventArgs e)
-        {
-            // hide ViewItemID column, also in header
-            e.Row.Cells [1].Visible = false;
-
-            // exclude header
-            if (e.Row.RowType == DataControlRowType.DataRow)
-            {
-                // find edit and delete linkbuttons
-                var linkDelete = e.Row.Cells [0].FindControl ("linkDelete") as LinkButton;
-                var linkEdit = e.Row.Cells [0].FindControl ("linkEdit") as LinkButton;
-
-                // set recordId to delete
-                linkEdit.CommandArgument = e.Row.Cells [1].Text;
-                linkDelete.CommandArgument = e.Row.Cells [1].Text;
-
-                // add confirmation dialog to delete link
-                linkDelete.Attributes.Add ("onClick", "javascript:return confirm('" + Localization.GetString ("DeleteItem") + "');");
-            }
-        }
-        #endregion
 	}
 }
