+++ R7.University/Controls/EditDocuments.ascx.control.cs
@@ -1,37 +1,25 @@
-﻿//
-// EditDocuments.ascx.control.cs
-//
-// Author:
-//       Roman M. Yagodin <roman.yagodin@gmail.com>
-//
-// Copyright (c) 2015 Roman M. Yagodin
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-using System;
+﻿using System.Web;
+using System.Web.UI;
+using System.Web.UI.WebControls;
+using System.Web.UI.HtmlControls;
+using DotNetNuke.UI.UserControls;
+using DotNetNuke.UI.WebControls;
+using DotNetNuke.Web.UI.WebControls;
 
-namespace R7.University
+namespace R7.University.Controls
 {
-    public class EditDocuments
+    public partial class EditDocuments
     {
-        public EditDocuments ()
-        {
-        }
+        protected GridView gridDocuments;
+        protected TextBox textDocumentTitle;
+        protected DropDownList comboDocumentType;
+        protected UrlControl urlDocumentUrl;
+        protected TextBox textDocumentSortIndex;
+        protected DnnDateTimePicker datetimeDocumentStartDate;
+        protected DnnDateTimePicker datetimeDocumentEndDate;
+        protected HiddenField hiddenDocumentItemID;
+        protected LinkButton buttonAddDocument;
+        protected LinkButton buttonUpdateDocument;
+        protected LinkButton buttonCancelEditDocument;
     }
 }
-
